Setup basic QA automation and update js-sdk to the currently latest
In order to test this quite outdated plugin against event more outdated versions of shopware 5,
starting with 5.3.0, an automated test is required.
In this commit there are a few tests that are introduced:
- code sniffer, to test the syntax
- phpstan for statical analysis against different shop versions
- phpmd, mess detector, to test against some bad practices
- php version compatibility check, checks against php 7.4 - 8.3

There is also a script written, that downloads required shop versions from
github and loads their composer dependencies.

Ref: S5P-114

+++ EnderecoShopware5Client.php
@@ -32,6 +32,9 @@ class EnderecoShopware5Client extends Plugin
         parent::build($container);
     }
 
+    /**
+     * @SuppressWarnings(PHPMD.UnusedFormalParameter)
+     */
     public function install(InstallContext $installContext)
     {
         $this->_addAttributes();
@@ -148,7 +151,9 @@ class EnderecoShopware5Client extends Plugin
         }
 
         $metaDataCache = Shopware()->Models()->getConfiguration()->getMetadataCacheImpl();
-        $metaDataCache->deleteAll();
+        if (method_exists($metaDataCache, 'deleteAll')) {
+            $metaDataCache->deleteAll();
+        }
         Shopware()->Models()->generateAttributeModels(['s_user_addresses_attributes','s_order_attributes']);
         $uninstallContext->scheduleClearCache(DeactivateContext::CACHE_LIST_ALL);
     }
@@ -335,7 +340,9 @@ class EnderecoShopware5Client extends Plugin
         }
 
         $metaDataCache = Shopware()->Models()->getConfiguration()->getMetadataCacheImpl();
-        $metaDataCache->deleteAll();
+        if (method_exists($metaDataCache, 'deleteAll')) {
+            $metaDataCache->deleteAll();
+        }
         Shopware()->Models()->generateAttributeModels(['s_user_addresses_attributes','s_order_attributes']);
     }
 }
